requirements.txt
python-telegram-bot==20.7
telethon==1.24.0
pyrogram==2.0.106
python-dotenv==1.0.0
psutil==5.9.0
colorama==0.4.6
requests==2.31.0
